graphiti-server/pyproject.toml
[project]
name = "graphiti-rest-server"
version = "1.0.0"
description = "Graphiti REST Server"
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.68.2",
    "graphiti-core>=0.21.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
graphiti-rest-server = "run:main"
